refactor machine learning model to base model
